% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{inputs}
\alias{inputs}
\alias{inputs<-}
\title{inputs}
\usage{
inputs(object)

inputs(object) <- value
}
\arguments{
\item{object}{A \code{snakemake} object.}

\item{value}{The value to assigin to inputs.}
}
\value{
The workflow inputs.
}
\description{
inputs

inputs
}
