% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWorkflow.R
\name{runWorkflow}
\alias{runWorkflow}
\title{run Workflow}
\usage{
runWorkflow(sobject, inputList, outdir = NULL, workdir = NULL, ...)
}
\arguments{
\item{sobject}{A \code{Snakemake} object.}

\item{inputList}{A list of inputs.}

\item{outdir}{The output directory.}

\item{workdir}{The \code{snakemake} work directory.}

\item{...}{More arguments from \code{snakemake_cmd}.}
}
\description{
To run workflow from \code{Snakemake} object and list of inputs.
}
