% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWorkflow.R
\name{runWorkflowBatch}
\alias{runWorkflowBatch}
\title{run Workflow in Batch}
\usage{
runWorkflowBatch(
  sobject,
  inputList,
  batch_by,
  outdir = NULL,
  workdir = NULL,
  BPPARAM = BatchtoolsParam(),
  ...
)
}
\arguments{
\item{sobject}{A \code{Snakemake} object.}

\item{inputList}{The list of inputs.}

\item{batch_by}{To run the workflow in batches by splitting the
input list.}

\item{outdir}{The output directory.}

\item{workdir}{The \code{Snakemake} directory if not the same as the
output directory.}

\item{BPPARAM}{The options for \code{BiocParallelParam}.}

\item{...}{More arguments for \code{snakemake_cmd}.}
}
\description{
run Workflow in Batch
}
