% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Snakemake-class}
\alias{Snakemake-class}
\alias{Snakemake}
\title{Snakemake class}
\usage{
Snakemake(workflow, inputs, outputs, rules)
}
\arguments{
\item{workflow}{The path of the Snakemake file.}

\item{inputs}{The list of inputs from configure file.}

\item{outputs}{The list of outputs.}

\item{rules}{The list of rules.}
}
\value{
A \code{snakemake} object
}
\description{
Snakemake class
}
