% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snakemake.R
\name{snakemake_cmd}
\alias{snakemake_cmd}
\title{snakemake cmd}
\usage{
snakemake_cmd(
  snakefile,
  configfile = NULL,
  workdir = NULL,
  cores = 1,
  Args = character(),
  showLog = TRUE
)
}
\arguments{
\item{snakefile}{The path to the Snakefile.}

\item{configfile}{The list of config files.}

\item{workdir}{The path to working directory.}

\item{cores}{The number of CPU cores to use.}

\item{Args}{More arguments for \code{snakemake}. Details can be found
at:
https://snakemake.readthedocs.io/en/stable/executing/cli.html}

\item{showLog}{Whether to print stdout and stderr.}
}
\value{
The logs.
}
\description{
Run workflow by snakemake cmd
}
